Moving kafka common classes to org.apache.pinot.plugin.stream.kafka
